refactor: shorten file names
