+++ frontend/app/admin/ingest/page.tsx
@@ -1,12 +1,1 @@
-import type { Metadata } from "next";
-
-import { AdminIngestView } from "@components/admin/ingest-view";
-
-export const metadata: Metadata = {
-  title: "Ingestion",
-  description: "Trigger and monitor listing ingestion runs",
-};
-
-const AdminIngestPage = () => <AdminIngestView />;
-
-export default AdminIngestPage;
+﻿
