+++ frontend/app/listings/page.tsx
@@ -1,4 +1,4 @@
-"use client";
+﻿"use client";
 
 import { useEffect, useState } from "react";
 import { usePathname } from "next/navigation";
@@ -6,7 +6,7 @@ import { usePathname } from "next/navigation";
 import { ButtonPrimary } from "@components/ui/button-primary";
 import { detailToModel } from "@components/property/listing-model";
 import type { PropertyModel } from "@typings/property";
-import { PropertyCard, PropertyCardSkeleton, PROPERTY_GRID } from "@components/property/property-card";
+import { PropertyCard, PropertyCardSkeleton, PROPERTY_GRID } from "@components/property/card";
 import { Heading2 } from "@components/ui/heading2";
 import { brand } from "@config/brand";
 import { listingsService } from "@services/listings";
