+++ frontend/components/landing/featured-listings/index.tsx
@@ -1,11 +1,11 @@
-"use client";
+﻿"use client";
 
 import { useEffect, useState } from "react";
 
 import { brand } from "@config/brand";
 import { detailToModel } from "@components/property/listing-model";
 import type { PropertyModel } from "@typings/property";
-import { PropertyCard, PropertyCardSkeleton, PROPERTY_GRID } from "@components/property/property-card";
+import { PropertyCard, PropertyCardSkeleton, PROPERTY_GRID } from "@components/property/card";
 import { ButtonSecondary } from "@components/ui/button-secondary";
 import { Heading2 } from "@components/ui/heading2";
 import { listingsService } from "@services/listings";
