+++ frontend/components/property/card.tsx
