+++ frontend/components/property/gallery.tsx
