+++ frontend/components/property/grid-with-map.tsx
